@@ -5,27 +5,26 @@ homepage: https://github.com/pulyaevskiy/node-interop
 author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 
 environment:
-  sdk: '>=2.9.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  analyzer: '>=0.36.4 <0.41.0'
-  bazel_worker: ^0.1.18
-  build: '>=1.3.0 <2.0.0'
-  build_config: '>=0.3.0 <0.5.0'
-  build_modules: ^2.11.3
+  bazel_worker: ^1.0.0
+  build: ^2.0.0
+  build_config: ">=0.3.0 <0.5.0"
+  build_modules: ^3.0.6
   collection: ^1.0.0
-  crypto: ^2.0.0
-  logging: ^0.11.3
+  crypto: ^3.0.0
+  logging: ^1.0.0
   meta: ^1.1.0
-  node_preamble: ^1.4.8
+  node_preamble: ^2.0.0
   path: ^1.4.2
   pool: ^1.3.6
-  scratch_space: ^0.0.2
+  scratch_space: ^1.0.0
 
 dev_dependencies:
-  build_test: ^0.10.0
+  build_test: ^2.0.0
   pedantic: ^1.0.0
-  test: ^1.0.0
+  test: ^1.16.8
   a:
     path: test/fixtures/a
   b:
