@@ -52,10 +52,9 @@ class NodeEntrypointBuilder implements Builder {
       {this.dart2JsArgs = const ['--no-sound-null-safety']});
 
   factory NodeEntrypointBuilder.fromOptions(BuilderOptions options) {
-    validateOptions(
-        options.config, _supportedOptions, 'build_node_compilers|entrypoint',
+    validateOptions(options.config, _supportedOptions, 'build_node_compilers|entrypoint',
         deprecatedOptions: _deprecatedOptions);
-    var compilerOption = options.config[_compiler] as String ?? 'dartdevc';
+    var compilerOption = options.config[_compiler] as String? ?? 'dartdevc';
     WebCompiler compiler;
     switch (compilerOption) {
       case 'dartdevc':
@@ -65,18 +64,16 @@ class NodeEntrypointBuilder implements Builder {
         compiler = WebCompiler.Dart2Js;
         break;
       default:
-        throw ArgumentError.value(compilerOption, _compiler,
-            'Only `dartdevc` and `dart2js` are supported.');
+        throw ArgumentError.value(
+            compilerOption, _compiler, 'Only `dartdevc` and `dart2js` are supported.');
     }
 
     if (options.config[_dart2jsArgs] is! List) {
-      throw ArgumentError.value(options.config[_dart2jsArgs], _dart2jsArgs,
-          'Expected a list for $_dart2jsArgs.');
+      throw ArgumentError.value(
+          options.config[_dart2jsArgs], _dart2jsArgs, 'Expected a list for $_dart2jsArgs.');
     }
-    var dart2JsArgs = (options.config[_dart2jsArgs] as List)
-            ?.map((arg) => '$arg')
-            ?.toList() ??
-        const <String>[];
+    var dart2JsArgs =
+        (options.config[_dart2jsArgs] as List?)?.map((arg) => '$arg').toList() ?? const <String>[];
 
     return NodeEntrypointBuilder(compiler, dart2JsArgs: dart2JsArgs);
   }
@@ -115,8 +112,7 @@ Future<bool> _isAppEntryPoint(AssetId dartId, AssetReader reader) async {
   // Skip reporting errors here, dartdevc will report them later with nicer
   // formatting.
   // ignore: deprecated_member_use
-  var parsed = parseString(
-      content: await reader.readAsString(dartId), throwIfDiagnostics: false);
+  var parsed = parseString(content: await reader.readAsString(dartId), throwIfDiagnostics: false);
 
   // Allow two or fewer arguments so that entrypoints intended for use with
   // [spawnUri] get counted.
@@ -126,6 +122,6 @@ Future<bool> _isAppEntryPoint(AssetId dartId, AssetReader reader) async {
   return parsed.unit.declarations.any((node) {
     return node is FunctionDeclaration &&
         node.name.name == 'main' &&
-        node.functionExpression.parameters.parameters.length <= 2;
+        node.functionExpression.parameters!.parameters.length <= 2;
   });
 }
